requirements.txt
RPi.GPIO==0.7.1
pika==1.3.1
adafruit-circuitpython-dht==3.7.7
board==1.0
adafruit-blinka==8.25.0
python-dotenv==1.0.0 